requirements.txt
Django==4.0.3
django-dotenv==1.4.2
djangorestframework==3.13.1
drf-yasg==1.20.0
python-decouple==3.6

